Move boilerplate code to own file
